Added Player, Added Controls
Hi Zoey! I added a spaceship player to the file and made it so the arrow keys would control it. I also made it so the player cannot go outside of the bounds of the map. I was having difficulty with the super class, so I removed the X and Y variables from the Graphic Object superclass. I couldn't figure out any way around this. Let me know if you have any questions!